+++ packages/testharness/src/host/index.js
@@ -3,9 +3,10 @@ import error from './error';
 import pageError from './pageError';
 import param from './param';
 import saveFile from './saveFile';
+import sendAccessKey from './sendAccessKey';
 import sendKeys from './sendKeys';
 import sendShiftTab from './sendShiftTab';
 import sendTab from './sendTab';
 import snapshot from './snapshot';
 
-export { done, error, pageError, param, saveFile, sendKeys, sendShiftTab, sendTab, snapshot };
+export { done, error, pageError, param, saveFile, sendAccessKey, sendKeys, sendShiftTab, sendTab, snapshot };