+++ packages/testharness/src/host/sendAccessKey.js
@@ -1,0 +1,10 @@
+import { post } from '../jobs';
+
+export default async function sendAccessKey(key) {
+  return await post({
+    payload: {
+      key
+    },
+    type: 'send access key'
+  });
+}